# frontend/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "高考志愿填报智能体",
    "description": "基于位次法的可解释、可回测志愿推荐系统（数据层当前为模拟数据，严禁真实填报）。所有响应含 data / evidence / warnings 三段。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/meta/provinces/{province}/rule": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "单省规则（批次级）",
        "operationId": "get_rule_api_v1_meta_provinces__province__rule_get",
        "parameters": [
          {
            "name": "province",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Province"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ProvinceMeta_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meta/provinces/{province}/subject-coverage": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "选考组合的可报专业覆盖率",
        "description": "首屏 Step 2 的\"实时可报专业覆盖率\"（真实统计，非估算）。\n\n返回的每个数字都来自当年招生计划；``subject_pool.origin`` 标明科目池是\n官方规则（RULE）还是由招生计划反推（DATA_DERIVED）。",
        "operationId": "get_subject_coverage_api_v1_meta_provinces__province__subject_coverage_get",
        "parameters": [
          {
            "name": "province",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Province"
            }
          },
          {
            "name": "subjects",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "逗号分隔的选考科目，如 物理,化学,生物",
              "default": "",
              "title": "Subjects"
            },
            "description": "逗号分隔的选考科目，如 物理,化学,生物"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_SubjectCoveragePayload_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/colleges/search": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "院校检索",
        "operationId": "search_colleges_api_v1_colleges_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "院校名模糊匹配",
              "title": "Q"
            },
            "description": "院校名模糊匹配"
          },
          {
            "name": "province",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "院校所在省",
              "title": "Province"
            },
            "description": "院校所在省"
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "层次标签，如 985 / 211 / 双一流",
              "title": "Level"
            },
            "description": "层次标签，如 985 / 211 / 双一流"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_CollegeSearchItem__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/majors/search": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "专业检索",
        "operationId": "search_majors_api_v1_majors_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "专业名模糊匹配",
              "title": "Q"
            },
            "description": "专业名模糊匹配"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "门类，如 工学",
              "title": "Category"
            },
            "description": "门类，如 工学"
          },
          {
            "name": "discipline",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "专业类，如 计算机类",
              "title": "Discipline"
            },
            "description": "专业类，如 计算机类"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 300,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_MajorSearchItem__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/units/{unit_id}/history": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "投档单位逐年历史",
        "operationId": "unit_history_api_v1_units__unit_id__history_get",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          },
          {
            "name": "years",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "default": 3,
              "title": "Years"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_UnitHistoryPayload_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CollegeSearchItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "province": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Province"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "level_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Level Tags"
          },
          "college_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "College Type"
          },
          "affiliation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affiliation"
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public",
            "default": true
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CollegeSearchItem"
      },
...
      "Envelope_ProvinceMeta_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProvinceMeta"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[ProvinceMeta]"
      },
...
      "Envelope_SubjectCoveragePayload_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SubjectCoveragePayload"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[SubjectCoveragePayload]"
      },
...
      "Envelope_UnitHistoryPayload_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UnitHistoryPayload"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[UnitHistoryPayload]"
      },
...
      "Envelope_list_CollegeSearchItem__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CollegeSearchItem"
            },
            "type": "array",
            "title": "Data"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[list[CollegeSearchItem]]"
      },
      "Envelope_list_MajorSearchItem__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MajorSearchItem"
            },
            "type": "array",
            "title": "Data"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[list[MajorSearchItem]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MajorSearchItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "discipline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discipline"
          },
          "degree": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degree"
          },
          "duration": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "MajorSearchItem"
      },
...
      "ProvinceMeta": {
        "properties": {
          "province": {
            "type": "string",
            "title": "Province"
          },
          "main_batch_code": {
            "type": "string",
            "title": "Main Batch Code"
          },
          "current_year": {
            "type": "integer",
            "title": "Current Year",
            "default": 2026
          },
          "subject_pool": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubjectPoolMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "batches": {
            "items": {
              "$ref": "#/components/schemas/BatchMeta"
            },
            "type": "array",
            "title": "Batches"
          },
          "requires_banner": {
            "type": "boolean",
            "title": "Requires Banner",
            "default": false
          },
          "has_caution": {
            "type": "boolean",
            "title": "Has Caution",
            "default": false
          },
          "source_problems": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Problems"
          }
        },
        "type": "object",
        "required": [
          "province",
          "main_batch_code"
        ],
        "title": "ProvinceMeta",
        "description": "``GET /meta/provinces`` 的单省条目。"
      },
...
      "SubjectCoveragePayload": {
        "properties": {
          "province": {
            "type": "string",
            "title": "Province"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "batch_code": {
            "type": "string",
            "title": "Batch Code"
          },
          "subjects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subjects"
          },
          "total_units": {
            "type": "integer",
            "title": "Total Units",
            "default": 0
          },
          "matched_units": {
            "type": "integer",
            "title": "Matched Units",
            "default": 0
          },
          "coverage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage"
          },
          "subject_pool": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubjectPoolMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "by_subject": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Subject"
          },
          "unmatched_examples": {
            "items": {
              "$ref": "#/components/schemas/UnmatchedUnit"
            },
            "type": "array",
            "title": "Unmatched Examples"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url",
            "default": ""
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "province",
          "year",
          "batch_code"
        ],
        "title": "SubjectCoveragePayload",
        "description": "选考组合的可报专业覆盖率（真实统计，非估算）。"
      },
      "SubjectPoolMeta": {
        "properties": {
          "province": {
            "type": "string",
            "title": "Province"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "choose": {
            "type": "integer",
            "title": "Choose",
            "default": 3
          },
          "subjects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subjects"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url",
            "default": ""
          },
          "source_quote": {
            "type": "string",
            "title": "Source Quote",
            "default": ""
          },
          "verified_status": {
            "$ref": "#/components/schemas/VerifiedStatus",
            "default": "UNVERIFIED"
          },
          "verified_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Year"
          },
          "caveats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Caveats"
          },
          "origin": {
            "type": "string",
            "title": "Origin",
            "default": "RULE"
          },
          "requires_caution": {
            "type": "boolean",
            "title": "Requires Caution",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "province"
        ],
        "title": "SubjectPoolMeta",
        "description": "省份选考科目池（AGENTS.md §8.1 Step 2）。\n\n``origin=\"RULE\"`` = 官方来源的规则事实；``origin=\"DATA_DERIVED\"`` = 未核实到原文时\n由招生计划反推的降级视图（``requires_caution=True``，前端必须提示，不得当成官方科目池）。"
      },
...
      "UnitHistoryPayload": {
        "properties": {
          "unit_id": {
            "type": "string",
            "title": "Unit Id"
          },
          "unit_key": {
            "type": "string",
            "title": "Unit Key"
          },
          "unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdmissionUnit"
              },
              {
                "type": "null"
              }
            ]
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/UnitHistoryRecord"
            },
            "type": "array",
            "title": "Records"
          }
        },
        "type": "object",
        "required": [
          "unit_id",
          "unit_key"
        ],
        "title": "UnitHistoryPayload"
      },
...
      "UnmatchedUnit": {
        "properties": {
          "unit_id": {
            "type": "string",
            "title": "Unit Id"
          },
          "college_id": {
            "type": "string",
            "title": "College Id"
          },
          "major_name": {
            "type": "string",
            "title": "Major Name"
          },
          "requirement_mode": {
            "type": "string",
            "title": "Requirement Mode"
          },
          "requirement_subjects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requirement Subjects"
          }
        },
        "type": "object",
        "required": [
          "unit_id",
          "college_id",
          "major_name",
          "requirement_mode"
        ],
        "title": "UnmatchedUnit",
        "description": "覆盖率统计里\"因选考不匹配被排除\"的示例单位。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifiedStatus": {
        "type": "string",
        "enum": [
          "PRIMARY",
          "PRIMARY_GOV",
          "SECONDARY",
          "UNVERIFIED"
        ],
        "title": "VerifiedStatus",
        "description": "规则核实等级（DOMAIN_RULES.md §1.2 / DECISIONS.md ADR-005）。"
      },
...
    }
  }
}